LPD-38764 Fix ID format

+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/reader/BatchEngineImportTaskItemReaderUtil.java
@@ -260,7 +260,7 @@ public class BatchEngineImportTaskItemReaderUtil {
 			_setFieldValue(object, "externalReferenceCode", creatorErc);
 		}
 
-		int creatorId = MapUtil.getInteger(values, "id");
+		long creatorId = MapUtil.getLong(values, "id");
 
 		if (creatorId != 0) {
 			_setFieldValue(object, "id", creatorId);
